+++ src/blob/persistence/SqlBlobMetadataStore.ts
@@ -8,7 +8,8 @@ import {
   Options as SequelizeOptions,
   Sequelize,
   TEXT,
-  Transaction
+  Transaction,
+  WhereOptions
 } from "sequelize";
 
 import uuid from "uuid/v4";
@@ -646,34 +647,19 @@ export default class SqlBlobMetadataStore implements IBlobMetadataStore {
         transaction: t
       });
 
-      // TODO: GC blobs under deleting status
-      await BlobsModel.update(
-        {
-          deleting: literal("deleting + 1")
+      await this.deleteBlobFromSQL({
+          accountName: account,
+          containerName: container
         },
-        {
-          where: {
-            accountName: account,
-            containerName: container
-          },
-          transaction: t
-        }
+        t
       );
 
-      // TODO: GC blocks under deleting status
-      await BlocksModel.update(
-        {
-          deleting: literal("deleting + 1")
+      await this.deleteBlockFromSQL({
+          accountName: account,
+          containerName: container
         },
-        {
-          where: {
-            accountName: account,
-            containerName: container
-          },
-          transaction: t
-        }
+        t
       );
-      /* Transaction ends */
     });
   }
 
@@ -1509,6 +1495,7 @@ export default class SqlBlobMetadataStore implements IBlobMetadataStore {
           blobName: block.blobName,
           blockName: block.name,
           size: block.size,
+          deleting: 0,
           persistency: this.serializeModelValue(block.persistency)
         },
         { transaction: t }
@@ -1977,51 +1964,33 @@ export default class SqlBlobMetadataStore implements IBlobMetadataStore {
         if (count > 1) {
           throw StorageErrorFactory.getSnapshotsPresent(context.contextId!);
         } else {
-          await BlobsModel.update(
-            {
-              deleting: literal("deleting + 1")
+          await this.deleteBlobFromSQL({
+              accountName: account,
+              containerName: container,
+              blobName: blob
             },
-            {
-              where: {
-                accountName: account,
-                containerName: container,
-                blobName: blob
-              },
-              transaction: t
-            }
+            t
           );
 
-          await BlocksModel.update(
-            {
-              deleting: literal("deleting + 1")
+          await this.deleteBlockFromSQL({
+              accountName: account,
+              containerName: container,
+              blobName: blob
             },
-            {
-              where: {
-                accountName: account,
-                containerName: container,
-                blobName: blob
-              },
-              transaction: t
-            }
+            t
           );
         }
       }
 
       // Scenario: Delete one snapshot only
       if (!againstBaseBlob) {
-        await BlobsModel.update(
-          {
-            deleting: literal("deleting + 1")
-          },
-          {
-            where: {
+        await this.deleteBlobFromSQL({
               accountName: account,
               containerName: container,
               blobName: blob,
               snapshot: blobModel.snapshot
             },
-            transaction: t
-          }
+            t
         );
       }
 
@@ -2030,32 +1999,19 @@ export default class SqlBlobMetadataStore implements IBlobMetadataStore {
         againstBaseBlob &&
         options.deleteSnapshots === Models.DeleteSnapshotsOptionType.Include
       ) {
-        await BlobsModel.update(
-          {
-            deleting: literal("deleting + 1")
+        await this.deleteBlobFromSQL({
+            accountName: account,
+            containerName: container,
+            blobName: blob
           },
-          {
-            where: {
-              accountName: account,
-              containerName: container,
-              blobName: blob
-            },
-            transaction: t
-          }
+          t
         );
 
-        await BlocksModel.update(
-          {
-            deleting: literal("deleting + 1")
-          },
-          {
-            where: {
-              accountName: account,
-              containerName: container,
-              blobName: blob
-            },
-            transaction: t
-          }
+        await this.deleteBlockFromSQL({
+            accountName: account,
+            containerName: container,
+            blobName: blob
+          },t
         );
       }
 
@@ -2064,19 +2020,13 @@ export default class SqlBlobMetadataStore implements IBlobMetadataStore {
         againstBaseBlob &&
         options.deleteSnapshots === Models.DeleteSnapshotsOptionType.Only
       ) {
-        await BlobsModel.update(
-          {
-            deleting: literal("deleting + 1")
+        await this.deleteBlobFromSQL({
+            accountName: account,
+            containerName: container,
+            blobName: blob,
+            snapshot: { [Op.gt]: "" }
           },
-          {
-            where: {
-              accountName: account,
-              containerName: container,
-              blobName: blob,
-              snapshot: { [Op.gt]: "" }
-            },
-            transaction: t
-          }
+          t
         );
       }
     });
@@ -3545,6 +3495,62 @@ export default class SqlBlobMetadataStore implements IBlobMetadataStore {
       return Models.AccessTier.Cold;
     }
     return undefined;
+  }  
+
+  /**
+   * Delete blob from SQL database.
+   * For performance, we used to mark deleting+1, instead of really delete. But this take issue like #2563. So change to real delete.
+   *
+   * @private
+   * @param {WhereOptions<any>} where
+   * @param {Transaction} [t]
+   * @returns {Promise<void>}
+   * @memberof SqlBlobMetadataStore
+   */
+  private async deleteBlobFromSQL(where: WhereOptions<any>, t?: Transaction): Promise<void> {
+    await BlobsModel.destroy({
+      where,
+      transaction: t
+    });
+    
+    // // TODO: GC blobs under deleting status
+    // await BlobsModel.update(
+    //   {
+    //     deleting: literal("deleting + 1")
+    //   },
+    //   {
+    //     where,
+    //     transaction: t
+    //   }
+    // );
+  }
+
+    /**
+   * Delete block from SQL database.
+   * For performance, we used to mark deleting+1, instead of really delete. But this take issue like #2563. So change to real delete.
+   *
+   * @private
+   * @param {WhereOptions<any>} where
+   * @param {Transaction} [t]
+   * @returns {Promise<void>}
+   * @memberof SqlBlobMetadataStore
+   */
+  private async deleteBlockFromSQL(where: WhereOptions<any>, t?: Transaction): Promise<void> {
+     await BlocksModel.destroy({
+      where,
+      transaction: t
+    });
+
+    // TODO: GC blocks under deleting status
+    // await BlocksModel.update(
+    //   {
+    //     deleting: literal("deleting + 1")
+    //   },
+    //   {
+    //     where,
+    //     transaction: t
+    //   }
+    // );
   }
 
   /**